HOOK_BOOST_3.0/requirements.txt
discord.py==2.3.2
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0 